+++ src/constants/path.ts
@@ -1,0 +1,8 @@
+export const LOGIN_ROUTER_PATH = {
+  login: '/account/login',
+  join: '/account/join',
+  password: {
+    find: '/account/password/find',
+    reset: '/account/password/reset',
+  },
+};
